+++ src/interfaces/ApiException.ts
@@ -1,0 +1,5 @@
+export interface ApiException {
+    statusCode: number;
+    message?: string;
+    instance?: string;
+}
